Разделил методы по стандарту es2015
