+++ js/names.js
@@ -1,0 +1,1 @@
+export const namesExamples = ["Саша","Оля","Андрей","Вика","Коля","Дима","Юля","Леша","Арсений","Валя", "Костя", "Вероника", "Алина"];
