+++ js/utils.js
@@ -1,0 +1,10 @@
+export const MAX_LENGHT_ARRAY_PHOTOS = 25;
+
+export const getRandomInt = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+export const getRandomArrayElement = (elements) => elements[getRandomInt(0, elements.length - 1)];
